adding javafx support
