+++ src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module app.app {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+
+    opens app to javafx.fxml;
+    exports app;
+}
